Cambio tipo de dato de numero documento a Long
Se hace el cambio del tipo de dato de numero documento de integer a Long, para evitar conflictos durante creacion de nuevos usuarios

+++ src/main/java/com/datacenter/mallaturnos/application/Dto/Usuario/CrearUsuarioDto.java
@@ -11,7 +11,7 @@ public class CrearUsuarioDto {
 
     private String nombre;
     private String tipoDocumento;
-    private Integer numeroDocumento;
+    private Long numeroDocumento;
     private Integer contrasena;
     private Long rolId;
     private Long cargoId;
